Fixed CORS and ported completely to Flask
