--- README.md
@@ -1,2 +1,0 @@
-# online_compiler
-Online python compiler
