--- compiler_api/app.py
@@ -1,24 +1,0 @@
-from flask import Flask,request,jsonify
-app = Flask(__name__)
-
-@app.route('/')
-def index():
-  return 'Server Works!'
-  
-@app.route('/greet')
-def say_hello():
-  return 'Hello from Server'
-
-@app.route('/api/', methods=["POST"])
-def main_interface():
-    response = request.get_json()
-    x=(response['message'])
-    y=eval(x)
-    response['message']=y
-    return jsonify(response)
-
-@app.after_request
-def add_headers(response):
-    response.headers.add('Access-Control-Allow-Origin', '*')
-    response.headers.add('Access-Control-Allow-Headers', 'Content-Type,Authorization')
-    return response
