+++ templates/index.html
@@ -4,8 +4,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/3.0.3/normalize.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/milligram/1.3.0/milligram.css">
     <link rel="stylesheet" href="https://milligram.io/styles/main.css">
-    <link rel="stylesheet" href="/css/dark.css">
-    <script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
+    <link rel="stylesheet" href="/static/css/dark.css">
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>OnlineCompiler</title>
@@ -17,6 +16,6 @@
 <textarea  class='dataField' id="message"></textarea>
 <button id="send">Send</button>
 <textarea id='compiler' class='compiler'></textarea>
-<script src="js/script.js"></script>
+<script src="/static/js/script.js"></script>
 </body>
 </html>
